fix(core,executor,planner,server,bench): widen NUMERIC backing to i128 with checked overflow (22003), closing silent truncation/overflow
Value::Decimal is now i128-backed (~38 significant digits) instead of i64.
NUMERIC literals and computed values up to i128 are represented exactly;
values beyond i128 raise SQLSTATE 22003 (numeric_value_out_of_range)
instead of silently truncating or wrapping.

- core: Decimal.value i64->i128; parse_decimal_text / decode_pg_numeric_binary
  no longer narrow to i64; write_decimal_text and encode_pg_numeric_binary
  widened; money parse range-checks the i128->i64 (cents) narrowing.
- executor: arithmetic (+,-,*,/,%,abs) computes and stores i128 with checked
  ops -> 22003 on overflow; MUL uses i128 checked_mul; compare/sort/value_key
  rescale and compare i128; casts and round/floor/ceil/trunc widened; SUM/AVG
  decimal accumulators carry i128 and error cleanly on i128 overflow.
- row codec / batches: heap decimal encoding is base-10000 (width-agnostic,
  on-disk format unchanged); decimal columns now materialise as decimal text
  in batches (lossless) instead of a fixed-width Int64 column that truncated
  beyond i64 (column_builder, decode_builders, build_batch, project_expr,
  filter_op, xml_table_scan); grouped vectorised SUM excludes Decimal so it
  uses the i128 row path.
- planner: out-of-i64 integer/decimal literals now bind exactly to i128
  NUMERIC (matching PG) rather than the i64 stopgap; beyond i128 -> 22003.
- server: wire text/binary, COPY, and export read decimals from the text
  column / i128 value.
- bench: TPC-H loader decimal encode widened to i128.

WAL replay round-trips unchanged (stores opaque heap tuple bytes). Decimal
index keys and decimal->int casts remain unsupported (pre-existing, 0A000).

+++ crates/ultrasql-bench/src/tpch/load/tests/sidecars_a.rs
@@ -67,7 +67,7 @@ fn tpch_q1_direct_sidecar_rejects_discount_factor_overflow() {
         payload.extend_from_slice(&value.to_le_bytes());
     }
     for value in [100_i64, 1_000, i64::MIN, 0] {
-        encode_direct_decimal(&mut payload, value, 2, 0).expect("decimal payload");
+        encode_direct_decimal(&mut payload, i128::from(value), 2, 0).expect("decimal payload");
     }
     payload.extend_from_slice(&1_u32.to_le_bytes());
     payload.push(b'N');
@@ -93,7 +93,7 @@ fn tpch_q6_direct_sidecar_rejects_revenue_overflow() {
         payload.extend_from_slice(&value.to_le_bytes());
     }
     for value in [1_00_i64, 10_000, 5, 0] {
-        encode_direct_decimal(&mut payload, value, 2, 0).expect("decimal payload");
+        encode_direct_decimal(&mut payload, i128::from(value), 2, 0).expect("decimal payload");
     }
     payload.extend_from_slice(&1_u32.to_le_bytes());
     payload.push(b'N');
@@ -155,7 +155,7 @@ fn tpch_q3_sidecar_reads_lineitem_payload_without_resplitting_text() {
         payload.extend_from_slice(&value.to_le_bytes());
     }
     for value in [100_i64, 10_000, 5, 0] {
-        encode_direct_decimal(&mut payload, value, 2, 0).expect("decimal payload");
+        encode_direct_decimal(&mut payload, i128::from(value), 2, 0).expect("decimal payload");
     }
     payload.extend_from_slice(&1_u32.to_le_bytes());
     payload.push(b'N');
@@ -191,7 +191,7 @@ fn tpch_q3_sidecar_rejects_discount_factor_overflow() {
         payload.extend_from_slice(&value.to_le_bytes());
     }
     for value in [100_i64, 10_000, i64::MIN, 0] {
-        encode_direct_decimal(&mut payload, value, 2, 0).expect("decimal payload");
+        encode_direct_decimal(&mut payload, i128::from(value), 2, 0).expect("decimal payload");
     }
     payload.extend_from_slice(&1_u32.to_le_bytes());
     payload.push(b'N');
@@ -219,7 +219,7 @@ fn tpch_q4_sidecar_counts_priority_when_lineitem_commits_before_receipt() {
         payload.extend_from_slice(&value.to_le_bytes());
     }
     for value in [100_i64, 10_000, 5, 0] {
-        encode_direct_decimal(&mut payload, value, 2, 0).expect("decimal payload");
+        encode_direct_decimal(&mut payload, i128::from(value), 2, 0).expect("decimal payload");
     }
     payload.extend_from_slice(&1_u32.to_le_bytes());
     payload.push(b'N');
@@ -268,7 +268,7 @@ fn tpch_q5_sidecar_sums_asia_revenue_for_matching_customer_supplier_nation() {
         payload.extend_from_slice(&value.to_le_bytes());
     }
     for value in [100_i64, 10_000, 5, 0] {
-        encode_direct_decimal(&mut payload, value, 2, 0).expect("decimal payload");
+        encode_direct_decimal(&mut payload, i128::from(value), 2, 0).expect("decimal payload");
     }
     payload.extend_from_slice(&1_u32.to_le_bytes());
     payload.push(b'N');
@@ -315,7 +315,7 @@ fn tpch_q5_sidecar_rejects_discount_factor_overflow() {
         payload.extend_from_slice(&value.to_le_bytes());
     }
     for value in [100_i64, 10_000, i64::MIN, 0] {
-        encode_direct_decimal(&mut payload, value, 2, 0).expect("decimal payload");
+        encode_direct_decimal(&mut payload, i128::from(value), 2, 0).expect("decimal payload");
     }
 
     let err = state
@@ -356,7 +356,7 @@ fn tpch_q7_sidecar_sums_france_germany_revenue_by_year() {
         payload.extend_from_slice(&value.to_le_bytes());
     }
     for value in [100_i64, 10_000, 5, 0] {
-        encode_direct_decimal(&mut payload, value, 2, 0).expect("decimal payload");
+        encode_direct_decimal(&mut payload, i128::from(value), 2, 0).expect("decimal payload");
     }
     payload.extend_from_slice(&1_u32.to_le_bytes());
     payload.push(b'N');
@@ -407,7 +407,7 @@ fn tpch_q7_sidecar_rejects_discount_factor_overflow() {
         payload.extend_from_slice(&value.to_le_bytes());
     }
     for value in [100_i64, 10_000, i64::MIN, 0] {
-        encode_direct_decimal(&mut payload, value, 2, 0).expect("decimal payload");
+        encode_direct_decimal(&mut payload, i128::from(value), 2, 0).expect("decimal payload");
     }
     payload.extend_from_slice(&1_u32.to_le_bytes());
     payload.push(b'N');
@@ -457,7 +457,7 @@ fn tpch_q8_sidecar_computes_brazil_market_share_by_year() {
         payload.extend_from_slice(&value.to_le_bytes());
     }
     for value in [100_i64, 10_000, 5, 0] {
-        encode_direct_decimal(&mut payload, value, 2, 0).expect("decimal payload");
+        encode_direct_decimal(&mut payload, i128::from(value), 2, 0).expect("decimal payload");
     }
 
     state
@@ -505,7 +505,7 @@ fn tpch_q8_sidecar_rejects_discount_factor_overflow() {
         payload.extend_from_slice(&value.to_le_bytes());
     }
     for value in [100_i64, 10_000, i64::MIN, 0] {
-        encode_direct_decimal(&mut payload, value, 2, 0).expect("decimal payload");
+        encode_direct_decimal(&mut payload, i128::from(value), 2, 0).expect("decimal payload");
     }
 
     let err = state
@@ -546,7 +546,7 @@ fn tpch_q9_sidecar_computes_green_part_profit_by_nation_year() {
         payload.extend_from_slice(&value.to_le_bytes());
     }
     for value in [100_i64, 10_000, 5, 0] {
-        encode_direct_decimal(&mut payload, value, 2, 0).expect("decimal payload");
+        encode_direct_decimal(&mut payload, i128::from(value), 2, 0).expect("decimal payload");
     }
 
     state
@@ -591,7 +591,7 @@ fn tpch_q9_sidecar_rejects_discount_factor_overflow() {
         payload.extend_from_slice(&value.to_le_bytes());
     }
     for value in [100_i64, 10_000, i64::MIN] {
-        encode_direct_decimal(&mut payload, value, 2, 0).expect("decimal payload");
+        encode_direct_decimal(&mut payload, i128::from(value), 2, 0).expect("decimal payload");
     }
 
     let err = state
